+++ CHANGELOG.md
@@ -17,7 +17,7 @@ blrec：`/blrec`修改为`/webhook/blrec`
 5. 增加审核后发送弹幕统计功能
 6. 为弹幕转换增加黑名单功能 -->
 
-1. 新增 `danmakuFactory路径`、`端口号`、`host`配置，webhook服务器自启动，默认为18010端口
+1. 新增 `danmakuFactory路径`、`port`、`host`配置，webhook服务器自启动，默认为18010端口
 2. 新增 `/webhook/bililiverecorder` 、`/webhook/blrec`、`/webhook/custom` webhook端点
 3. 优化工具页的文件选择栏&部分组件UI优化
 4. 暗色模式支持